Add files via upload
An updated version of my contribution to the code. Since idk how to push it through, I didn't merge it. This piece of code is only missing the login function and then the program will be complete. A notable change is that due to encountering issues with the writing and reading of files, I made the files open and close within a function to ensure maximum reliability.